// src/app/shared/enums/cuisine-type.enum.ts
export enum CuisineType {
  
  FAST_FOOD = 'FAST_FOOD',
  FINE_DINING = 'FINE_DINING',
  BISTRO = 'BISTRO',
  BRASSERIE = 'BRASSERIE',
  STREET_FOOD = 'STREET_FOOD',
  FAMILY = 'FAMILY',
  FRENCH = 'FRENCH',
  ITALIAN = 'ITALIAN',
  ASIAN = 'ASIAN',
  JAPANESE = 'JAPANESE',
  MEXICAN = 'MEXICAN',
  INDIAN = 'INDIAN',
  MEDITERRANEAN = 'MEDITERRANEAN',
  OTHER = 'OTHER'
}